+++ test/build.js
@@ -3,7 +3,7 @@ require("esbuild").build({
   entryPoints: ["./app.js"],
   bundle: true,
   outfile: "./out.js",
-  plugins: [require("..").style()],
+  plugins: [require("..").inlineSass()],
   target: 'es5',
   sourcemap: true
 }).catch(() => process.exit(1))